xls2network/requirements.txt
et-xmlfile==1.0.1
jdcal==1.3
openpyxl==2.4.9
pytest==3.2.3
ruamel.yaml<=0.15
